perf(config): lazy-load ApprovalFlowsTab to keep Phase 2 out of main chunk
Extracted the Approval Flows tab body into its own module and swapped
the static import in ConfigPanel for React.lazy + Suspense.
Previously WORKFLOW_TEMPLATES, useSavedWorkflows, and the tab render
tree were pulled into the main chunk every time ConfigPanel shipped,
costing ~3 kB gzipped whether or not anyone opened the tab.

Now Phase 2 adds only +0.24 kB gzip to the main chunk (well under
the plan's 2 kB budget). The full builder — validator, layout
engine, canvas, inspector, picker, simulator, modal shell, templates,
persistence hook — lives in two lazy chunks totalling 17.83 kB gzip
that download only when the owner opens the tab.

Audit doc updated with before/after measurements.
